Skip force termination of processes on first pass, in case file is in the process of being uploaded (MCS-1771).

+++ deploy_files/mess/monitor_process.sh
@@ -9,6 +9,8 @@ scene_file_basename=$(basename "$scene_file")
 
 # TA1 run command
 ta1_run_cmd="python src/script_mess_clean.py scenes/${scene_file_basename}"
+# Skip termination steps on first pass that history file is found, in case file is currently being uploaded
+first_run=true
 
 while true; do
     # check if TA1 process is running by looking for their run command
@@ -25,25 +27,31 @@ while true; do
         # shellcheck disable=SC2154
         if compgen -G "${eval_dir}/SCENE_HISTORY/${filename}*.json" > /dev/null
         then
-            echo "monitor_process.sh: History file for ${filename} exists, terminate process."
-            echo "monitor_process.sh: Terminate child unity process first"
-            child_processes=$(pgrep -P "$python_process")
-            for cpid in $child_processes;
-            do
-                echo "$cpid"
-                if ps -p "$cpid" > /dev/null
-                then
-                    kill -15 "$cpid"
-                fi
-            done
+            if [ "$first_run" = true ];
+            then
+                echo "monitor_process.sh: History file exists for ${filename}, skipping process termination on first pass in case file is currently being uploaded."
+                first_run=false
+            else
+                echo "monitor_process.sh: History file for ${filename} exists, terminate process."
+                echo "monitor_process.sh: Terminate child unity process first"
+                child_processes=$(pgrep -P "$python_process")
+                for cpid in $child_processes;
+                do
+                    echo "$cpid"
+                    if ps -p "$cpid" > /dev/null
+                    then
+                        kill -15 "$cpid"
+                    fi
+                done
 
-            echo "monitor_process.sh: Terminate main process now"
-            pkill -f "${ta1_run_cmd}"
-            sleep 20
-            echo "monitor_process.sh: exiting for ${scene_file}"
-            exit 1
+                echo "monitor_process.sh: Terminate main process now"
+                pkill -f "${ta1_run_cmd}"
+                sleep 20
+                echo "monitor_process.sh: exiting for ${scene_file}"
+                exit 1
+            fi
         fi
     fi
-    echo "monitor_process.sh: Sleeping for an hour to wait for output for ${scene_file}"
-    sleep 3600
+    echo "monitor_process.sh: Sleeping for half an hour to wait for output for ${scene_file}"
+    sleep 1800
 done
